fix: Resolve golangci-lint errors
- Fix errcheck errors in session/tracker.go
- Fix errcheck errors in memory/repository.go
- Fix errcheck errors in commands.go (Scanln, config.Save)
- Fix errcheck errors in health.go (rows.Scan)
- Fix integer overflow in embeddings/generator.go
- Remove unused variable (extensionAssets) in db.go
- Remove unused variable (globalGraph) in mcp/handlers.go
- Fix unparam error (unused query parameter) in mcp/handlers.go
- Fix formatting in tui/tutorial.go (import order)
- Fix deprecated help.NewModel() to help.New()
- Restore SetGraph function for API compatibility

+++ internal/embeddings/generator.go
@@ -163,7 +163,7 @@ func (g *TFIDFEmbeddingGenerator) generateEmbedding(tokens []string) []float32 {
 			if dim < 0 {
 				dim = -dim
 			}
-			embedding[dim] += score*float32((hash>>uint32(j%32))&1)*2 - 1
+			embedding[dim] += score*float32(uint32(hash>>uint32(j%32))&1)*2 - 1
 		}
 	}
